+++ src/modules/rtmp.js
@@ -13,10 +13,11 @@ var serverOptions = {
     'ping': 60,
     'ping_timeout': 30
   },
+
   'http': {
     'port': 8000,
     'allow_origin': '*',
-    'mediaroot': './media'
+    'mediaroot': 'src/public/videos'
   }
 };
 
@@ -45,16 +46,14 @@ if (config.streamRecording)
       hlsFlags: '[hls_time=2:hls_list_size=3:hls_flags=delete_segments]',
       dash: true,
       dashFlags: '[f=dash:window_size=3:extra_window_size=5]'
+    },
+    {
+      'app': 'live',
+      // 'ac': 'aac',
+      'mp4': true,
+      'mp4Flags': '[movflags=faststart]',
     }
     ]
-    // 'tasks': [
-    //   {
-    //     'app': 'live',
-    //     'ac': 'aac',
-    //     'mp4': true,
-    //     'mp4Flags': '[movflags=faststart]',
-    //   }
-    // ]
   }